# specs/002-roles-permissions-system/openapi.yaml
openapi: 3.0.3
info:
  title: WasteFlow Roles & Permissions API
  description: |
    Comprehensive roles and permissions management system for WasteFlow application.

    This API supports:
    - Multi-tenant role-based access control (RBAC)
    - Attribute-based access control (ABAC) policies
    - Consultant multi-tenant context switching
    - Temporary permission grants with auto-expiration
    - Comprehensive audit logging for regulatory compliance
    - Custom role creation for enterprise clients
    - Permission request workflows

    ## Authentication
    All endpoints require JWT Bearer authentication obtained via SPID/CIE login.
    The JWT contains user identity, tenant context, and SPID fiscal code for audit traceability.

    ## Multi-Tenancy
    - Schema-per-tenant architecture with Row-Level Security (RLS)
    - Tenant context extracted from JWT claims
    - Consultants can switch tenant context via dedicated endpoints
    - All operations scoped to current tenant unless explicitly cross-tenant (consultants only)

    ## Rate Limiting
    - Standard endpoints: 100 requests/minute per user
    - Permission check endpoints: 1000 requests/minute per user (high-frequency caching)
    - Audit log queries: 20 requests/minute per user (expensive operations)
    - Bulk operations: 10 requests/minute per user

    ## Compliance
    - All audit logs retained for 10+ years per D.Lgs. 152/2006
    - Immutable audit trail with cryptographic integrity
...

  version: 1.0.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /audit/permissions:
    get:
      tags: [Audit Logs]
      summary: Query permission audit logs
      description: |
        Returns immutable audit trail of permission checks and changes.
        Supports complex filtering for compliance reporting.
        Expensive operation - rate limited to 20 requests/minute.
        Results can be exported to CSV for regulatory submissions.
      operationId: queryPermissionAuditLog
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by specific user
        - name: action
          in: query
          schema:
            type: string
          description: Filter by action attempted (e.g., fir:create)
        - name: decision
          in: query
          schema:
            type: string
            enum: [ALLOW, DENY, all]
            default: all
          description: Filter by authorization decision
        - name: resourceType
          in: query
          schema:
            type: string
          description: Filter by resource type (fir, facility, report, etc.)
        - name: resourceId
          in: query
          schema:
            type: string
          description: Filter by specific resource ID
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
          description: Start of time range (ISO 8601)
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
          description: End of time range (ISO 8601)
        - name: fiscalCode
          in: query
          schema:
            type: string
          description: Filter by SPID fiscal code
      responses:
        '200':
          description: Audit logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      x-rate-limit: 20/minute
  /audit/permissions/export:
    post:
      tags: [Audit Logs]
      summary: Export audit logs to CSV
      description: |
        Generates CSV export of audit logs matching query criteria.
        Maximum 50,000 records per export.
        Async job - returns job ID for download link retrieval.
        Export retained for 7 days then auto-deleted.
      operationId: exportAuditLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditExportRequest'
      responses:
        '202':
          description: Export job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      x-rate-limit: 5/minute
  /audit/role-changes:
    get:
      tags: [Audit Logs]
      summary: Query role/permission change history
      description: |
        Returns audit trail of all role and permission modifications.
        Includes before/after snapshots for compliance reconstruction.
        Enables historical query: "Who could access X at time Y?"
      operationId: queryRoleChangeHistory
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: entityType
          in: query
          schema:
            type: string
            enum: [Role, UserRole, RolePermission, all]
            default: all
        - name: changeType
          in: query
          schema:
            type: string
            enum: [CREATE, UPDATE, DELETE, ASSIGN, REVOKE, all]
            default: all
        - name: changedBy
          in: query
          schema:
            type: string
            format: uuid
          description: User who made the change
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Role change history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoleChangeHistoryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      x-rate-limit: 20/minute
  /audit/permissions/reconstruct:
    post:
      tags: [Audit Logs]
      summary: Reconstruct historical permissions
      description: |
        Answers: "Who could access resource X at timestamp Y?"
        Reconstructs permission state from versioned role assignment history.
        Critical for compliance investigations and forensic analysis.
      operationId: reconstructHistoricalPermissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resourceType:
                  type: string
                  description: Type of resource (fir, facility, etc.)
                resourceId:
                  type: string
                  description: Specific resource ID
                timestamp:
                  type: string
                  format: date-time
                  description: Point in time to reconstruct (ISO 8601)
                action:
                  type: string
                  description: Action to check (create, read, update, delete, etc.)
              required: [resourceType, resourceId, timestamp, action]
      responses:
        '200':
          description: Historical permissions reconstructed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricalPermissionsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      x-rate-limit: 10/minute

  # ===========================
  # CONSULTANT MULTI-TENANT CONTEXT
  # ===========================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained via SPID/CIE authentication.
        Include in Authorization header: `Bearer <token>`

        JWT contains claims:
        - userId: User unique identifier
        - tenantId: Current tenant context
        - fiscalCode: SPID fiscal code for audit
        - roles: User's role IDs
        - permissions: Cached permission set
        - exp: Token expiration timestamp
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Items per page
...
  schemas:
...
    PermissionAuditLogEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userName:
          type: string
        fiscalCode:
          type: string
          description: SPID fiscal code for regulatory traceability
        action:
          type: string
          description: Action attempted
          example: fir:delete
        resourceType:
          type: string
          description: Type of resource
          example: fir
        resourceId:
          type: string
          format: uuid
          nullable: true
          description: Specific resource ID
        decision:
          type: string
          enum: [ALLOW, DENY]
        reason:
          type: string
          description: Reason for decision
        evaluatedPolicies:
          type: array
          items:
            type: string
          description: ABAC policies evaluated
        contextAttributes:
          type: object
          description: Context at time of check
          properties:
            ipAddress:
              type: string
            userAgent:
              type: string
            deviceType:
              type: string
            sessionId:
              type: string
              format: uuid
            facilityId:
              type: string
              format: uuid
        consultantContext:
          type: object
          nullable: true
          description: Consultant context if applicable
          properties:
            isConsultant:
              type: boolean
            consultantFirm:
              type: string
            actingForTenant:
              type: string
              format: uuid
        timestamp:
          type: string
          format: date-time
          description: Timestamp with millisecond precision
        evaluationTimeMs:
          type: number
          format: float
          description: Decision evaluation time in milliseconds
      required: [id, tenantId, userId, userName, fiscalCode, action, decision, reason, timestamp]
    PaginatedAuditLogResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PermissionAuditLogEntry'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        filterSummary:
          type: object
          description: Summary of applied filters
          properties:
            allowCount:
              type: integer
            denyCount:
              type: integer
            uniqueUsers:
              type: integer
            dateRange:
              type: object
              properties:
                from:
                  type: string
                  format: date-time
                to:
                  type: string
                  format: date-time
      required: [items, total, page, limit, totalPages]
    AuditExportRequest:
      type: object
      properties:
        filters:
          type: object
          description: Same filter parameters as query endpoint
          properties:
            userId:
              type: string
              format: uuid
            action:
              type: string
            decision:
              type: string
              enum: [ALLOW, DENY]
            resourceType:
              type: string
            resourceId:
              type: string
            dateFrom:
              type: string
              format: date-time
            dateTo:
              type: string
              format: date-time
            fiscalCode:
              type: string
        format:
          type: string
          enum: [csv, json]
          default: csv
        includeHeaders:
          type: boolean
          default: true
        maxRecords:
          type: integer
          minimum: 1
          maximum: 50000
          default: 10000
          description: Maximum records to export (max 50k)
      required: [filters]
    RoleChangeHistoryEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entityType:
          type: string
          enum: [Role, UserRole, RolePermission]
        entityId:
          type: string
          format: uuid
        changeType:
          type: string
          enum: [CREATE, UPDATE, DELETE, ASSIGN, REVOKE]
        changedBy:
          type: string
          format: uuid
        changedByName:
          type: string
        changedByFiscalCode:
          type: string
        reason:
          type: string
        beforeSnapshot:
          type: object
          description: Entity state before change (JSON)
        afterSnapshot:
          type: object
          description: Entity state after change (JSON)
        timestamp:
          type: string
          format: date-time
        affectedUsers:
          type: array
          items:
            type: string
            format: uuid
          description: Users affected by this change
      required: [id, entityType, entityId, changeType, changedBy, timestamp]
    PaginatedRoleChangeHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RoleChangeHistoryEntry'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, limit, totalPages]
    HistoricalPermissionsResponse:
      type: object
      properties:
        resourceType:
          type: string
        resourceId:
          type: string
        timestamp:
          type: string
          format: date-time
        action:
          type: string
        usersWithAccess:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
                format: uuid
              userName:
                type: string
              fiscalCode:
                type: string
              roles:
                type: array
                items:
                  type: string
              grantingPermissions:
                type: array
                items:
                  type: string
                description: Which specific permissions granted access
        reconstructionMethod:
          type: string
          description: How permissions were reconstructed
          example: Analyzed role assignments and permission changes from 2024-01-15 to query timestamp
        confidence:
          type: string
          enum: [HIGH, MEDIUM, LOW]
          description: Confidence level of reconstruction
      required: [resourceType, resourceId, timestamp, action, usersWithAccess]

    # ===========================
    # CONSULTANT SCHEMAS
    # ===========================
...
    AsyncJobResponse:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        jobType:
          type: string
        status:
          type: string
          enum: [QUEUED, PROCESSING, COMPLETED, FAILED]
        createdAt:
          type: string
          format: date-time
        estimatedCompletionTime:
          type: string
          format: date-time
        statusUrl:
          type: string
          description: URL to poll for job status
      required: [jobId, jobType, status, createdAt, statusUrl]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: PERMISSION_DENIED
        message:
          type: string
          description: Human-readable error message (Italian)
          example: Non hai il permesso necessario per eseguire questa azione
        statusCode:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error context
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint that generated error
      required: [error, message, statusCode, timestamp]
    ValidationErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Field that failed validation
        message:
          type: string
          description: Validation error message
        value:
          description: Invalid value provided
        constraint:
          type: string
          description: Validation constraint violated
      required: [field, message]
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Authentication required. Please login with SPID/CIE.
            statusCode: 401
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
    ForbiddenError:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: Non hai il permesso necessario per eseguire questa azione
            statusCode: 403
            details:
              requiredPermission: roles:create
              currentRole: OPERATOR
              contactAdministrator: admin@example.com
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  validationErrors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationErrorDetail'
          example:
            error: VALIDATION_ERROR
            message: Request validation failed
            statusCode: 400
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
            validationErrors:
              - field: name
                message: Role name must be at least 3 characters
                value: ab
                constraint: minLength
              - field: permissions
                message: At least one permission is required
                constraint: minItems
...